requirements_pip.txt
absl-py==1.3.0
aiohttp==3.8.3
annotated-types==0.6.0
astroid==2.15.8
async-timeout==4.0.2
attrs==22.1.0
blinker==1.4
bottleneck==1.3.5
brotli==1.0.9
brotlipy==0.7.0
certifi==2024.2.2
cffi==1.15.1
click==8.0.4
continuum==1.2.7
contourpy==1.0.5
cryptography==38.0.1
datasets==2.18.0
dill==0.3.8
dodgy==0.2.1
exceptiongroup==1.2.0
filelock==3.13.1
flake8==5.0.4
flake8-polyfill==1.0.2
frozenlist==1.3.3
fsspec==2023.10.0
gitdb==4.0.11
gitpython==3.1.42
grpcio==1.42.0
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.4
imagehash==4.3.1
imageio==2.34.0
importlib-metadata==4.11.3
iniconfig==2.0.0
isort==5.13.2
joblib==1.1.1
kiwisolver==1.4.4
lazy-loader==0.3
lazy-object-proxy==1.10.0
lightning==2.2.1
lightning-utilities==0.9.0
markdown==3.3.4
matplotlib==3.7.0
mccabe==0.7.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.16
mypy==1.9.0
mypy-extensions==1.0.0
networkx==3.2.1
numexpr==2.8.4
numpy==1.23.4
oauthlib==3.2.1
pandas==1.5.1
pep8-naming==0.10.0
pillow==9.2.0
pip==22.2.2
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
prospector==1.10.3
protobuf==3.20.1
psutil==5.9.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pycodestyle==2.9.1
pydantic==2.7.1
pydantic-core==2.18.2
pydocstyle==6.3.0
pyflakes==2.5.0
pyjwt==2.4.0
pylint==2.17.7
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.0.9
pyqt5-sip==12.11.0
pysocks==1.7.1
pytest==8.1.1
pytest-mock==3.12.0
pytz==2023.3.post1
pywavelets==1.5.0
pyyaml==6.0.1
requests==2.28.1
requirements-detector==1.2.2
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.1.3
scipy==1.9.3
semver==3.0.2
setoptconf-tmp==0.3.1
setuptools==65.5.0
sip==6.6.2
smmap==5.0.1
snowballstemmer==2.2.0
tensorboard  #==2.10.0
tensorboard-data-server #==0.6.1
tensorboard-plugin-wit #==1.8.1
tifffile==2024.2.12
tomli==2.0.1
tomlkit==0.12.4
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.65.0
typing-extensions==4.11.0
urllib3==1.26.18
wrapt==1.16.0
xxhash==3.4.1
yarl==1.8.1
zipp==3.8.0
